config/fstab.ci20: Mount system partition RO
This is necessary in order to fix first boot when SELinux
is in enforcing mode :

[   43.060063] type=1400 audit(188.170:19): avc: denied { write } for pid=391 comm="patchoat"
    name="boot.art" dev="ubifs" ino=1798 scontext=u:r:zygote:s0
    tcontext=u:object_r:system_file:s0 tclass=file permissive=1

The root cause of the issue is that patchoat is trying to make a
private mapping of the boot.art file in memory RW.

http://androidxref.com/6.0.1_r10/xref/art/patchoat/patchoat.cc#153

This should not be a problem, because private mappings should never be flushed back
down to disk. It seems that this type of mapping is behaving differently
if it was created upon a file which resides on a RW mounted partition.

Change-Id: I7c4da2940f6994816631665c2b5f03ab76926886
Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>