Fix UBIFS build
sudo command fails to access stdin for password when it is called
from a shell script, which is executed during Android build.
The error was not present in the earlier versions of Android
and was introduced after switching to ninja build system.

Error log is the following:
sudo: no tty present and no askpass program specified
sudo: no tty present and no askpass program specified
ninja: build stopped: subcommand failed.
make: *** [ninja_wrapper] Error 1

Because of this  convert_extfs_to_ubifs script call
has been moved to mksdcard script and system.ubifs.img
is being created if it does not exist or in case that
it has an older date compared to system.img.

Build procedure is simpler after this change when compared to
previous versions Android UBIFS build. It does not pause
the build and wait for password because of EXT4/UBIFS conversion.
Instead, conversion is being done right before burning
the image to sdcard and this is only done if needed.
Burning the image required root access even before
this change, so this step will not require additional
user interaction.

Change-Id: I066ee07ed4b219b2bb91467d1cf576b8a36a625b
Signed-off-by: Dragan Cecavac <dragan.cecavac@imgtec.com>
(cherry picked from commit 1c79cebf30d286868dff5c170d4b78942b63ac13)